Move usePostFields to wordpress/fields package
