--- packages/edit-site/src/components/post-fields/style.scss
@@ -1,3 +1,0 @@
-.components-popover.components-dropdown__content.dataforms-layouts-panel__field-dropdown {
-	z-index: z-index(".components-popover.components-dropdown__content.dataforms-layouts-panel__field-dropdown");
-}
