+++ packages/fields/src/fields/index.ts
@@ -8,3 +8,4 @@ export { default as statusField } from './status';
 export { default as commentStatusField } from './comment-status';
 export { default as dateField } from './date';
 export { default as authorField } from './author';
+export { default as usePostFields } from './use-post-fields';
